Add: - Contribution to ui.menus to define new link requirement action in the toolbar - Modification of the update/create algorithms in wizards - Add counting algorithm description in the preference page

+++ plugins/org.topcased.requirement.core/src/org/topcased/requirement/core/extensions/DefaultAttachmentPolicy.java
@@ -107,7 +107,7 @@ public class DefaultAttachmentPolicy implements IModelAttachmentPolicy
      * @see org.topcased.requirement.core.extensions.IModelAttachmentPolicy#unlinkRequirementModel(org.eclipse.emf.ecore.resource.Resource,
      *      org.eclipse.emf.ecore.resource.Resource)
      */
-    public void unlinkRequirementModel(Resource targetModel, Resource requirementModel)
+    public void unlinkRequirementModel(Resource targetModel, Resource requirementModel, boolean deleteRequirementModel)
     {
         if (getProperty(targetModel.getContents().get(0)) != null)
         {
@@ -123,7 +123,10 @@ public class DefaultAttachmentPolicy implements IModelAttachmentPolicy
             // unload and delete the requirement model from file system.
             if (modeler.getEditingDomain().getResourceSet().getResources().remove(requirementModel))
             {
-                RequirementUtils.deleteResource(requirementModel);
+                if (deleteRequirementModel)
+                {
+                    RequirementUtils.deleteResource(requirementModel);
+                }
             }
         }
     }
